+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [1.1.1][unreleased] - (in development)
-- TBD
+- Changed: Dependency updates
+- Changed: Add application header
+- Removed: Java 1.7 build
 
 ## Version [1.1.0] - 2016-06-21
 - Fixed: Do not generate omitted fields